drake-gym/drake_gym/examples/wandb/run-20260202_230401-uqzfcxgr/files/requirements.txt
rpds-py==0.17.1
fsspec==2024.10.0
tqdm==4.67.1
autoprop==4.1.0
dill==0.3.7
sentry-sdk==2.51.0
mdurl==0.1.2
idna==3.6
colorama==0.4.0
lockfile==0.12.2
bayesian-optimization==1.4.0
manipulation==2024.11.17
MarkupSafe==2.1.3
jedi==0.19.1
cssutils==2.9.0
gradescope-utils==0.5.0
cma==4.0.0
natsort==8.4.0
tinycss2==1.2.1
ipython==8.20.0
pexpect==4.9.0
sphinxcontrib-qthelp==1.0.7
gitdb==4.0.12
mdit-py-plugins==0.4.0
json5==0.9.14
Send2Trash==1.8.2
bleach==6.1.0
chardet==5.2.0
scikit-image==0.24.0
pytz==2023.3.post1
snowballstemmer==2.2.0
ruamel.yaml==0.18.5
empy==3.3.4
jupyter_client==8.6.0
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cufft-cu12==11.0.2.54
websocket-client==1.7.0
pydantic==2.12.5
rich==13.7.0
sphinx-autodoc-typehints==1.25.2
ipykernel==6.29.0
referencing==0.32.1
mpmath==1.3.0
grpcio==1.76.0
sphinxemoji==0.2.0
nvidia-nvtx-cu12==12.1.105
debugpy==1.8.0
pytest==9.0.2
networkx==3.4.2
nbimporter==0.3.4
CacheControl==0.13.1
lxml==5.1.0
more-itertools==10.5.0
jupyter_server==2.12.5
sympy==1.13.3
traitlets==5.14.1
click==8.3.1
ptyprocess==0.7.0
attrs==23.2.0
nvidia-cusolver-cu12==11.4.5.107
parso==0.8.3
python-dateutil==2.8.2
jupyter-events==0.9.0
beautifulsoup4==4.12.2
pyparsing==3.1.1
tornado==6.4
vhacdx==0.0.8.post1
psutil==5.9.8
filelock==3.13.1
certifi==2023.11.17
nbclient==0.9.0
tabulate==0.9.0
jupyterlab_widgets==3.0.13
Sphinx==7.2.6
sphinx-book-theme==1.1.0
pydot==2.0.0
soupsieve==2.5
meshio==5.3.5
PyQt5-sip==12.13.0
Werkzeug==3.1.5
trimesh==4.1.8
sniffio==1.3.0
ruamel.yaml.clib==0.2.8
iniconfig==2.3.0
sphinxcontrib-jsmath==1.0.1
charset-normalizer==3.3.2
accessible-pygments==0.0.4
taichi==1.7.0
executing==2.0.1
nvidia-cublas-cu12==12.1.3.1
ipywidgets==8.1.5
nvidia-cuda-runtime-cu12==12.1.105
wcwidth==0.2.13
shiboken2==5.15.2.1
jupyterlab==4.0.11
linkify-it-py==2.0.2
packaging==23.2
msgpack==1.0.7
Jinja2==3.1.3
Cython==3.0.11
widgetsnbextension==4.0.13
torch==2.3.1
dict2css==0.3.0.post1
tomli==2.0.1
mpld3==0.5.10
signature_dispatch==1.0.1
pluggy==1.6.0
cycler==0.12.1
importlib-metadata==7.0.1
overrides==7.4.0
annotated-types==0.7.0
jupyter_core==5.7.1
myst-parser==2.0.0
stable_baselines3==2.4.0
sphinx-tabs==3.4.4
apeye-core==1.1.4
kiwisolver==1.4.5
matplotlib-inline==0.1.6
scipy==1.14.1
Pygments==2.17.2
tensorboard-data-server==0.7.2
comm==0.2.1
cffi==1.16.0
domdf-python-tools==3.8.0.post2
asttokens==2.4.1
smmap==5.0.2
nvidia-cusparse-cu12==12.1.0.106
pycparser==2.21
drake==1.35.0
cloudpickle==2.2.1
async-lru==2.0.4
catkin-pkg==1.0.0
nvidia-nccl-cu12==2.20.5
markdown-it-py==3.0.0
Markdown==3.10.1
argon2-cffi-bindings==21.2.0
alabaster==0.7.16
wheel==0.42.0
pyglet==1.5.30
nbformat==5.9.2
wandb==0.24.1
pydantic_core==2.41.5
zipp==3.17.0
lazy_loader==0.4
jupyterlab_pygments==0.3.0
PySide2==5.15.2.1
apeye==1.4.1
numpy==1.26.3
imageio==2.36.1
directsearch==1.0
pure-eval==0.2.2
PyQt5==5.15.10
stack-data==0.6.3
platformdirs==4.1.0
pyzmq==24.0.1
sphinxcontrib-applehelp==1.0.8
webencodings==0.5.1
decorator==5.1.1
fastjsonschema==2.19.1
pip==26.0
jupyterlab_server==2.25.2
typing-inspection==0.4.2
pillow==10.2.0
contourpy==1.2.0
nbconvert==7.14.2
torchvision==0.18.1
sphinx-jinja2-compat==0.2.0.post1
tensorboard==2.20.0
joblib==1.4.2
nevergrad==1.0.8
lark==1.2.2
arrow==1.3.0
fonttools==4.47.2
python-fcl==0.7.0.6
GitPython==3.1.46
nvidia-curand-cu12==10.3.2.106
jsonschema-specifications==2023.12.1
nest-asyncio==1.5.9
jsonpointer==2.4
anyio==4.2.0
docutils==0.18.1
types-python-dateutil==2.8.19.20240106
sphinx-toolbox==3.5.0
defusedxml==0.7.1
jupyter-lsp==2.2.2
nvidia-cudnn-cu12==8.9.2.26
nvidia-nvjitlink-cu12==12.6.85
six==1.16.0
matplotlib==3.8.2
sphinxcontrib-serializinghtml==1.1.10
imagesize==1.4.1
argon2-cffi==23.1.0
gymnasium==1.0.0
triton==2.3.1
rfc3339-validator==0.1.4
isoduration==20.11.0
manifold3d==3.0.0
notebook_shim==0.2.3
pyserial==3.5
exceptiongroup==1.2.0
fqdn==1.5.1
timeout-decorator==0.5.0
sphinx-prompt==1.8.0
protobuf==6.33.5
pydata-sphinx-theme==0.15.1
notebook==7.0.7
setuptools==69.0.2
PyQt5-Qt5==5.15.2
autodocsumm==0.2.12
tzdata==2024.2
jupyter_server_terminals==0.5.1
scikit-learn==1.6.0
pandas==2.2.3
uc-micro-py==1.0.2
typeguard==3.0.2
html5lib==1.1
prometheus-client==0.19.0
terminado==0.18.0
uri-template==1.3.0
typing_extensions==4.15.0
pandocfilters==1.5.1
sphinxcontrib-htmlhelp==2.0.5
glooey==0.3.6
webcolors==1.13
requests==2.31.0
urllib3==2.1.0
rfc3986-validator==0.1.1
threadpoolctl==3.5.0
PyYAML==6.0.1
tifffile==2024.9.20
vecrec==0.3.1
absl-py==2.4.0
python-json-logger==2.0.7
Farama-Notifications==0.0.4
prompt-toolkit==3.0.43
jsonschema==4.21.0
sphinxcontrib-devhelp==1.0.6
mistune==3.0.2
Babel==2.14.0
nvidia-cuda-nvrtc-cu12==12.1.105
